Improve ParametricBodies initialization
I redefined `ParametricBodies,scale` so it can be calculated for mappings that extrude a 2D curve into 3D. However, this requires an input vector `x₀`, whose magnitude is irrelevant but whose `ndims=length(x₀)` can't generally be inferred!

I've added `ndims` as an optional input and we can set `ndims=NurbsCurve{N}` with a fallback to `ndims=2`, but this still needs to be supplied manually for extruded geometries.